rework eye + face synchronization logic
See https://github.com/elucideye/drishti/issues/669 for montage illustrating updated face/frame + eye synchronization behavior.

Since the optimized pipeline has a build in latency == 2, if the user requests a history of N, we configure the main video frame FIFO to be N + latency.   So for N=3 we would have a fifo.length() == 3+2 == 5.  In this optimized pipeline mode, previously, the first two frames would be returned without eye pairs.  This PR adds an ignoreLatestFramesInMonitor member variable to FaceFinder (and public FaceTracker counterpart) to control whether or not the leading frames will be included in the user requested N frames or not.  Given the config above, If ignoreLatestFramesInMonitor==true, then if the user requests 3 frames it will be assumed that they are only interested in the frames with paired metadata, so the first 2 frames will not be returned in the vector.  The user can set ignoreLatestFramesInMonitor=false to return to the old behavior, in which case the leading two frames will be returned even though they have no associated metadata or eye textures.  The ignoreLatestFramesInMonitor variable has no impact if the unoptimzed pipeline is being used, since it has no built-in latency.

Summary:

* rework face/frame + eye + metadata synchronization
* c++ formatting (reduce some longer protypes)
* in Facefinder::init() the latency member initializatio is initialized first and we rely on this for FIFO initialization (in particular, we allocate the full frame FIFO to be $((latency + history)) if we are using the optimized pipeline
* add FaceFinder::Impl::ignoreLatestFramesInMonitor (see texture above) and set to true by default
* remove unused